04-study-flashcards/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
mysql-connector-python>=9.0,<10.0
httpx==0.26.0
pytest==7.4.4
